LATEST UPDATE!!!

+++ app/Http/Controllers/PostsController.php
@@ -7,6 +7,9 @@ use App\Post;
 
 class PostsController extends Controller
 {
+    public function __construct(){
+        $this->middleware('auth', ['except'=> ['index', 'show']]);
+    }
     /**
      * Display a listing of the resource.
      *
@@ -38,13 +41,25 @@ class PostsController extends Controller
     {
         $this->validate($request, [
             'title' => 'required',
-            'body' => 'required'
+            'body' => 'required',
+            'cover_image' => 'image|nullable|max:1999'
         ]);
+
+        if($request->hasFile('cover_image')){
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME );
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            $filenameToStore = $filename.'_'.time().'.'.$extension;
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $filenameToStore);
+        } else {
+            $filenameToStore = 'no_image.jpg';
+        }
 
         $post = new Post();
         $post->title = $request->input('title');
         $post->body = $request->input('body');
         $post->user_id = auth()->user()->id;
+        $post->cover_image = $filenameToStore;
         $post->save();
 
         return redirect('/posts')->with('success','Post Successfuly Created!!!');
@@ -71,6 +86,11 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error', 'Unauthorized Page!!!');
+        }
+
         return view('posts.edit')->with('post',$post);
     }
 
@@ -105,6 +125,11 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error', 'Unauthorized Page!!!');
+        }
+
         $post->delete();
 
         return redirect('/posts')->with('success', 'Post Successfully Deleted!!!');
